requirements.txt
python-dotenv==1.0.0
Requests==2.31.0
python-telegram-bot==20.4